+++ CHANGELOG.rst
@@ -1,6 +1,7 @@
 1.4.0
 -----
 
+* Added :class:`.Slugify` computation (#466)
 * Added ``slugify_names`` argument to :meth:`Table.rename`. (#466)
 * Added :meth:`utils.slugify` to standardize a sequence of strings. (#466)
 * :meth:`Table.print_table` now ensures decimal points line up if numbers have different precision. (#553)